requirements.txt
streamlit
streamlit-mic-recorder
chromadb
sentence-transformers
PyMuPDF
google-generativeai
SpeechRecognition
pydub
pytest
tqdm
python-dotenv 